greatly cleaned up code in all modules for better readability, moved fetch functions to api manager for single responsibility principle
