+++ api/database.json
@@ -122,6 +122,43 @@
       "isWaves": true,
       "waterTemp": "90",
       "id": 8
+    },
+    {
+      "userId": 1,
+      "date": "2021-09-21",
+      "location": "asfsdaf",
+      "diveSite": "asdfasdf",
+      "freshOrSalt": "Salt",
+      "depth": 55,
+      "time": 55,
+      "comments": "raefasd",
+      "isDrift": true,
+      "isDry": true,
+      "isFFM": true,
+      "isN32": true,
+      "isCurrent": true,
+      "isShore": true,
+      "isSurge": true,
+      "waterTemp": "86",
+      "id": 9
+    },
+    {
+      "userId": 1,
+      "date": "2021-09-22",
+      "location": "USVI",
+      "diveSite": "sdfgsdfg",
+      "freshOrSalt": "Salt",
+      "depth": 55,
+      "time": 65,
+      "comments": "sdfgsfdg",
+      "isDeep": true,
+      "isDrift": true,
+      "isFFM": true,
+      "isNav": true,
+      "isBoat": true,
+      "isCurrent": true,
+      "waterTemp": "52",
+      "id": 10
     }
   ],
   "gear": [
@@ -234,12 +271,13 @@
       "tank": "TO RENT",
       "gloves": "",
       "hood": "7mm",
+      "camera": "Insta360",
       "id": 3
     },
     {
       "userId": 1,
       "name": "asdfasedf",
-      "bcd": "sadfasdf",
+      "bcd": "SHERWOOD",
       "regulator": "uytfg",
       "octopus": "ituyfi",
       "mask": "yfiyt",
@@ -288,60 +326,6 @@
       "towel": true,
       "waterBottle": true,
       "id": 4
-    },
-    {
-      "userId": 1,
-      "name": "Title",
-      "bcd": "",
-      "regulator": "",
-      "octopus": "",
-      "mask": "",
-      "fins": "",
-      "boots": "",
-      "console": "",
-      "computer": "",
-      "exposureSuit": "",
-      "weight": 0,
-      "tank": "",
-      "gloves": "",
-      "hood": "",
-      "batteries": "",
-      "camera": "",
-      "cameraAcc": "",
-      "chargers": "",
-      "o2Analyzer": "",
-      "power": "",
-      "strobe": "",
-      "torch": "",
-      "torchBackup": "",
-      "finStrap": false,
-      "maskBackup": false,
-      "maskStrap": false,
-      "mouthpiece": false,
-      "oRings": false,
-      "zipTies": false,
-      "itinerary": false,
-      "maps": false,
-      "passport": false,
-      "logbook": false,
-      "planeTicket": false,
-      "travelIns": false,
-      "vaxRec": false,
-      "audSignal": "",
-      "firstAid": false,
-      "knife": "",
-      "meds": "",
-      "o2Kit": false,
-      "sunscreen": "",
-      "vizSignal": "",
-      "wreckReel": "",
-      "dryBag": false,
-      "defog": false,
-      "sunglasses": false,
-      "toolKit": false,
-      "towel": false,
-      "waterBottle": false,
-      "id": 5
     }
   ],
   "certCards": [
