+++ src/components/diveLog/Conditions.js
@@ -1,26 +1,28 @@
 import React from "react";
 
-
 export const Conditions = ({ diveCopy, setDive }) => {
     return (<>
         <h2 className="diveNumber">Dive & Water Conditions</h2>
         <section className="specialties">
+
             <div className="specialtyColumn">
                 <ul className="specialtyList">
                     <li>
-                        <input type="checkbox" value="false" className="logCheckbox" onChange={(event) => {
-                            const copy = { ...diveCopy }
-                            copy.isBoat = event.target.checked
-                            setDive(copy)
-                        }} />
+                        <input type="checkbox" value="false" className="logCheckbox"
+                            onChange={(event) => {
+                                const copy = { ...diveCopy }
+                                copy.isBoat = event.target.checked
+                                setDive(copy)
+                            }} />
                         <label className="checkboxLabel">Boat</label>
                     </li>
                     <li>
-                        <input type="checkbox" value="false" className="logCheckbox" onChange={(event) => {
-                            const copy = { ...diveCopy }
-                            copy.isSurge = event.target.checked
-                            setDive(copy)
-                        }} />
+                        <input type="checkbox" value="false" className="logCheckbox"
+                            onChange={(event) => {
+                                const copy = { ...diveCopy }
+                                copy.isSurge = event.target.checked
+                                setDive(copy)
+                            }} />
                         <label className="checkboxLabel">Surge</label>
                     </li>
                 </ul>
@@ -29,19 +31,21 @@ export const Conditions = ({ diveCopy, setDive }) => {
             <div className="specialtyColumn">
                 <ul className="specialtyList">
                     <li>
-                        <input type="checkbox" className="logCheckbox" onChange={(event) => {
-                            const copy = { ...diveCopy }
-                            copy.isCurrent = event.target.checked
-                            setDive(copy)
-                        }} />
+                        <input type="checkbox" className="logCheckbox"
+                            onChange={(event) => {
+                                const copy = { ...diveCopy }
+                                copy.isCurrent = event.target.checked
+                                setDive(copy)
+                            }} />
                         <label className="checkboxLabel">Current/Drift</label>
                     </li>
                     <li>
-                        <input type="checkbox" className="logCheckbox" onChange={(event) => {
-                            const copy = { ...diveCopy }
-                            copy.isDry = event.target.checked
-                            setDive(copy)
-                        }} />
+                        <input type="checkbox" className="logCheckbox"
+                            onChange={(event) => {
+                                const copy = { ...diveCopy }
+                                copy.isDry = event.target.checked
+                                setDive(copy)
+                            }} />
                         <label className="checkboxLabel">Waves</label>
                     </li>
                 </ul>
@@ -50,11 +54,12 @@ export const Conditions = ({ diveCopy, setDive }) => {
             <div className="specialtyColumn">
                 <ul className="specialtyList">
                     <li>
-                        <input type="checkbox" className="logCheckbox" onChange={(event) => {
-                            const copy = { ...diveCopy }
-                            copy.isShore = event.target.checked
-                            setDive(copy)
-                        }} />
+                        <input type="checkbox" className="logCheckbox"
+                            onChange={(event) => {
+                                const copy = { ...diveCopy }
+                                copy.isShore = event.target.checked
+                                setDive(copy)
+                            }} />
                         <label className="checkboxLabel">Shore</label>
                     </li>
                 </ul>
@@ -63,11 +68,12 @@ export const Conditions = ({ diveCopy, setDive }) => {
 
         <fieldset>
             <label className="">Water Temperature</label>
-            <input type="number" width="20px" onChange={(event) => {
-                const copy = { ...diveCopy }
-                copy.waterTemp = event.target.value
-                setDive(copy)
-            }} />°F
+            <input type="number" width="20px"
+                onChange={(event) => {
+                    const copy = { ...diveCopy }
+                    copy.waterTemp = event.target.value
+                    setDive(copy)
+                }} />°F
         </fieldset>
     </>
     )
