+++ src/components/diveLog/EditDive.js
@@ -1,22 +1,14 @@
 import React, { useEffect, useState } from "react"
 import { useHistory, useParams } from "react-router";
-import { getCurrentDive } from "../application/ApiManager";
+import { getCurrentDive, saveDive } from "../application/ApiManager";
 import { EditConditions } from "./EditConditions";
 import { EditSpecialties } from "./EditSpecialties";
 
 export const EditDive = () => {
     const history = useHistory()
     const { diveId } = useParams()
-
     const [currentDive, setDive] = useState({
-        userId: parseInt(localStorage.getItem('si_user')),
-        date: '',
-        location: ``,
-        diveSite: '',
-        freshOrSalt: '',
-        depth: 0,
-        time: 0,
-        comments: ''
+        userId: parseInt(localStorage.getItem('si_user'))
     })
 
     useEffect(
@@ -28,55 +20,6 @@ export const EditDive = () => {
         },
         [diveId]
     )
-
-    const saveDive = (event) => {
-        event.preventDefault()
-        const updateDive = {
-            userId: currentDive.userId,
-            date: currentDive.date,
-            location: currentDive.location,
-            diveSite: currentDive.diveSite,
-            freshOrSalt: currentDive.freshOrSalt,
-            depth: currentDive.depth,
-            time: currentDive.time,
-            comments: currentDive.comments,
-            isAltitude: currentDive.isAltitude,
-            isCave: currentDive.isCave,
-            isDeep: currentDive.isDeep,
-            isDrift: currentDive.isDrift,
-            isDry: currentDive.isDry,
-            isFFM: currentDive.isFFM,
-            isNav: currentDive.isNav,
-            isNight: currentDive.isNight,
-            isN32: currentDive.isN32,
-            isN36: currentDive.isN36,
-            isRebreather: currentDive.isRebreather,
-            isSearch: currentDive.isSearch,
-            isSidemount: currentDive.isSidemount,
-            isWreck: currentDive.isWreck,
-            isBoat: currentDive.isBoat,
-            isCurrent: currentDive.isCurrent,
-            isShore: currentDive.isShore,
-            isSurge: currentDive.isSurge,
-            isWaves: currentDive.isWaves,
-            waterTemp: currentDive.waterTemp,
-        }
-
-        const fetchOption = {
-            method: "PUT",
-            headers: {
-                "Content-type": "application/json",
-            },
-            body: JSON.stringify(updateDive)
-        }
-
-        return fetch(`http://localhost:8088/dives/${diveId}`, fetchOption)
-            .then(
-                () => {
-                    history.push("/divelog")
-                }
-            )
-    }
 
     return (<>
         <article className='diveLogArticle'>
@@ -90,72 +33,85 @@ export const EditDive = () => {
                 <h2 className='diveNumber'>Dive {currentDive.id}</h2>
                 <fieldset>
                     <label for="date">Date </label>
-                    <input type="date" value={currentDive.date} className='date' onChange={(event) => {
-                        const copy = { ...currentDive }
-                        copy.date = event.target.value
-                        setDive(copy)
-                    }} />
+                    <input type="date" value={currentDive.date} className='date'
+                        onChange={(event) => {
+                            const copy = { ...currentDive }
+                            copy.date = event.target.value
+                            setDive(copy)
+                        }} />
                 </fieldset>
                 <fieldset>
                     <label for="location">Location</label>
-                    <input type="location" value={currentDive.location} onChange={(event) => {
-                        const copy = { ...currentDive }
-                        copy.location = event.target.value
-                        setDive(copy)
-                    }} />
+                    <input type="location" value={currentDive.location}
+                        onChange={(event) => {
+                            const copy = { ...currentDive }
+                            copy.location = event.target.value
+                            setDive(copy)
+                        }} />
                 </fieldset>
                 <fieldset>
                     <label for="diveSite">Dive Site</label>
-                    <input type="diveSite" value={currentDive.diveSite} onChange={(event) => {
-                        const copy = { ...currentDive }
-                        copy.diveSite = event.target.value
-                        setDive(copy)
-                    }} />
+                    <input type="diveSite" value={currentDive.diveSite}
+                        onChange={(event) => {
+                            const copy = { ...currentDive }
+                            copy.diveSite = event.target.value
+                            setDive(copy)
+                        }} />
                 </fieldset>
                 <div className="fresh-salt-radios">
                     <input type="radio" value="Fresh"
-                        checked={currentDive.freshOrSalt === 'Fresh' ? true : false} onChange={(event) => {
+                        checked={currentDive.freshOrSalt === 'Fresh' ? true : false}
+                        onChange={(event) => {
                             const copy = { ...currentDive }
                             copy.freshOrSalt = event.target.value
                             setDive(copy)
                         }} />
                     <label>Fresh</label>
-                    <input type="radio" value="Salt" checked={currentDive.freshOrSalt === 'Salt' ? true : false} onChange={(event) => {
-                        const copy = { ...currentDive }
-                        copy.freshOrSalt = event.target.value
-                        setDive(copy)
-                    }} /><label>Salt</label>
+                    <input type="radio" value="Salt"
+                        checked={currentDive.freshOrSalt === 'Salt' ? true : false}
+                        onChange={(event) => {
+                            const copy = { ...currentDive }
+                            copy.freshOrSalt = event.target.value
+                            setDive(copy)
+                        }} /><label>Salt</label>
                 </div>
                 <fieldset>
                     <label for="depth">Depth</label>
-                    <input type="number" value={currentDive.depth} onChange={(event) => {
-                        const copy = { ...currentDive }
-                        copy.depth = parseInt(event.target.value)
-                        setDive(copy)
-                    }} />
+                    <input type="number" value={currentDive.depth}
+                        onChange={(event) => {
+                            const copy = { ...currentDive }
+                            copy.depth = parseInt(event.target.value)
+                            setDive(copy)
+                        }} />
                 </fieldset>
                 <fieldset>
                     <label for="time">Time</label>
-                    <input type="number" value={currentDive.time} onChange={(event) => {
-                        const copy = { ...currentDive }
-                        copy.time = parseInt(event.target.value)
-                        setDive(copy)
-                    }} />
+                    <input type="number" value={currentDive.time}
+                        onChange={(event) => {
+                            const copy = { ...currentDive }
+                            copy.time = parseInt(event.target.value)
+                            setDive(copy)
+                        }} />
                 </fieldset>
                 <fieldset>
                     <label for="comments">Comments</label>
-                    <textarea type="comments" value={currentDive.comments} className="comments-box" rows="4" onChange={(event) => {
-                        const copy = { ...currentDive }
-                        copy.comments = event.target.value
-                        setDive(copy)
-                    }} />
+                    <textarea type="comments" value={currentDive.comments}
+                        className="comments-box" rows="4"
+                        onChange={(event) => {
+                            const copy = { ...currentDive }
+                            copy.comments = event.target.value
+                            setDive(copy)
+                        }} />
                 </fieldset>
 
-                    <EditConditions currentDive={currentDive} setDive={setDive}/>
-                    <EditSpecialties currentDive={currentDive} setDive={setDive}/>
+                <EditConditions currentDive={currentDive} setDive={setDive} />
+                <EditSpecialties currentDive={currentDive} setDive={setDive} />
 
                 <fieldset className="submit-cancel-buttons">
-                    <button className="submitButton" type="submit" onClick={saveDive}>
+                    <button className="submitButton" type="submit"
+                        onClick={(event) => saveDive(event, currentDive, diveId)
+                            .then(history.push("/divelog"))
+                        }>
                         Save Changes
                     </button>
                     <button className="cancelButton" onClick={() => {
