+++ src/components/gear/CreateGear.js
@@ -1,5 +1,6 @@
 import React, { useState } from "react";
 import { useHistory } from "react-router";
+import { submitGear } from "../application/ApiManager";
 import "../diveLog/LogNewDive.css";
 import { Backups } from "./additonalGear/Backups";
 import { Documents } from "./additonalGear/Documents";
@@ -12,139 +13,7 @@ import { EssentialGear } from "./EssentialGear";
 
 export const CreateGearSet = () => {
     const history = useHistory()
-
-    const [gearCopy, setNewGear] = useState({
-        //Essential
-        userId: parseInt(localStorage.getItem('si_user')),
-        name: '',
-        bcd: '',
-        regulator: '',
-        octopus: '',
-        mask: '',
-        fins: '',
-        boots: '',
-        console: '',
-        computer: '',
-        exposureSuit: '',
-        weight: 0,
-        tank: '',
-        // Exposure
-        gloves: '',
-        hood: '',
-        // Tech
-        batteries: '',
-        camera: '',
-        cameraAcc: '',
-        chargers: '',
-        o2Analyzer: '',
-        power: '',
-        strobe: '',
-        torch: '',
-        torchBackup: '',
-        // Backups
-        finStrap: false,
-        maskBackup: false,
-        maskStrap: false,
-        mouthpiece: false,
-        oRings: false,
-        zipTies: false,
-        // Documents
-        itinerary: false,
-        maps: false,
-        passport: false,
-        logbook: false,
-        planeTicket: false,
-        travelIns: false,
-        vaxRec: false,
-        // Safety
-        audSignal: '',
-        firstAid: false,
-        knife: '',
-        meds: '',
-        o2Kit: false,
-        sunscreen: '',
-        vizSignal: '',
-        wreckReel: '',
-        // Misc
-        dryBag: false,
-        defog: false,
-        sunglasses: false,
-        toolKit: false,
-        towel: false,
-        waterBottle: false
-    })
-
-    const submitGear = (event) => {
-        event.preventDefault()
-        const newGear = {
-            userId: gearCopy.userId,
-            name: gearCopy.name,
-            bcd: gearCopy.bcd,
-            regulator: gearCopy.regulator,
-            octopus: gearCopy.octopus,
-            mask: gearCopy.mask,
-            fins: gearCopy.fins,
-            boots: gearCopy.boots,
-            console: gearCopy.console,
-            computer: gearCopy.computer,
-            exposureSuit: gearCopy.exposureSuit,
-            weight: parseInt(gearCopy.weight),
-            tank: gearCopy.tank,
-            gloves: gearCopy.gloves,
-            hood: gearCopy.hood,
-            batteries: gearCopy.batteries,
-            camera: gearCopy.camera,
-            cameraAcc: gearCopy.cameraAcc,
-            chargers: gearCopy.chargers,
-            o2Analyzer: gearCopy.o2Analyzer,
-            power: gearCopy.power,
-            strobe: gearCopy.strobe,
-            torch: gearCopy.torch,
-            torchBackup: gearCopy.torchBackup,
-            finStrap: gearCopy.finStrap,
-            maskBackup: gearCopy.maskBackup,
-            maskStrap: gearCopy.maskStrap,
-            mouthpiece: gearCopy.mouthpiece,
-            oRings: gearCopy.oRings,
-            zipTies: gearCopy.zipTies,
-            itinerary: gearCopy.itinerary,
-            maps: gearCopy.maps,
-            passport: gearCopy.passport,
-            logbook: gearCopy.logbook,
-            planeTicket: gearCopy.planeTicket,
-            travelIns: gearCopy.travelIns,
-            vaxRec: gearCopy.vaxRec,
-            audSignal: gearCopy.audSignal,
-            firstAid: gearCopy.firstAid,
-            knife: gearCopy.knife,
-            meds: gearCopy.meds,
-            o2Kit: gearCopy.o2Kit,
-            sunscreen: gearCopy.sunscreen,
-            vizSignal: gearCopy.vizSignal,
-            wreckReel: gearCopy.wreckReel,
-            dryBag: gearCopy.dryBag,
-            defog: gearCopy.defog,
-            sunglasses: gearCopy.sunglasses,
-            toolKit: gearCopy.toolKit,
-            towel: gearCopy.towel,
-            waterBottle: gearCopy.waterBottle
-        }
-
-        const fetchOption = {
-            method: "POST",
-            headers: {
-                "Content-type": "application/json",
-            },
-            body: JSON.stringify(newGear)
-        }
-
-        return fetch("http://localhost:8088/gear", fetchOption)
-            .then(
-                () => {
-                    history.push("/gear")
-                }
-            )
-    }
+    const [gearCopy, setNewGear] = useState({ userId: parseInt(localStorage.getItem('si_user')), })
 
     return (<>
         <article className='diveLogArticle'>
@@ -172,7 +41,10 @@ export const CreateGearSet = () => {
                 <Misc gearCopy={gearCopy} setNewGear={setNewGear} />
 
                 <fieldset className="submit-cancel-buttons">
-                    <button className="submitButton" type="submit" onClick={submitGear}>
+                    <button className="submitButton" type="submit"
+                        onClick={(event) => submitGear(event, gearCopy)
+                            .then(history.push("/gear"))
+                        }>
                         Submit
                     </button>
                     <button className="cancelButton" onClick={() => {
