+++ src/components/gear/EssentialGear.js
@@ -2,7 +2,6 @@ import React, { useState } from "react";
 
 export const EssentialGear = ({ gearCopy, setNewGear }) => {
     const [toggle, setToggle] = useState(false)
-
     const toggleList = () => toggle !== true ? true : false
     const toggled = toggleList()
