+++ src/components/gear/Gear.js
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from "react"
 import { Link } from "react-router-dom"
-import { getMyGear } from "../application/ApiManager"
+import { deleteGear, getMyGear } from "../application/ApiManager"
 import "./Gear.css"
 import { GearList } from "./GearList"
 
@@ -17,30 +17,20 @@ export const Gear = () => {
         []
     )
 
-    const deleteGear = (id) => {
-        fetch(`http://localhost:8088/gear/${id}`, {
-            method: "DELETE"
-        })
-            .then(() => {
-                getMyGear()
-                    .then(gear => setGear(gear))
-            }, []
-            )
-    }
+    return (
+        <>
+            <article className='gearArticle'>
+                <h2 className='gearTitle'>Gear</h2>
+                <section className="quip-button">
+                    <h2 className="quip">Proper dive planning includes ensuring you have all the gear you need for the dive conditions.</h2>
+                    <Link to="/addgear"><button className="logDive">
+                        Create Gear Set<br />
+                        🤿🧳
+                    </button></Link>
+                </section>
+            </article>
 
-    return (<>
-        <article className='gearArticle'>
-            <h2 className='gearTitle'>Gear</h2>
-            <section className="quip-button">
-                <h2 className="quip">Proper dive planning includes ensuring you have all the gear you need for the dive conditions.</h2>
-                <Link to="/addgear"><button className="logDive">
-                    Create Gear Set<br />
-                    🤿🧳
-                </button></Link>
-            </section>
-        </article>
-
-        <GearList gear={gear} deleteGear={deleteGear} />
-
-    </>)
+            <GearList gear={gear} setGear={setGear} deleteGear={deleteGear} />
+        </>
+    )
 }
