+++ src/components/gear/additonalGear/Safety.js
@@ -2,7 +2,6 @@ import React, { useState } from "react";
 
 export const Safety = ({ gearCopy, setNewGear }) => {
     const [toggle, setToggle] = useState(false)
-
     const toggleList = () => toggle !== true ? true : false
     const toggled = toggleList()
