+++ src/components/gear/editGear/EditEssentials.js
@@ -2,7 +2,6 @@ import React, { useState } from "react";
 
 export const EditEssentials = ({ currentGear, updateGear }) => {
     const [toggle, setToggle] = useState(false)
-
     const toggleList = () => toggle !== true ? true : false
     const toggled = toggleList()
