+++ src/components/gear/editGear/EditExposure.js
@@ -2,7 +2,6 @@ import React, { useState } from "react";
 
 export const EditExposure = ({ currentGear, updateGear }) => {
     const [toggle, setToggle] = useState(false)
-
     const toggleList = () => toggle !== true ? true : false
     const toggled = toggleList()
 
@@ -31,7 +30,6 @@ export const EditExposure = ({ currentGear, updateGear }) => {
                             updateGear(copy)
                         }} />
                     </fieldset>
-
                 </>
             }
         </>
