+++ src/components/gear/editGear/EditSafety.js
@@ -1,8 +1,7 @@
 import React, { useState } from "react";
 
-export const EditSafety = ({currentGear, updateGear}) => {
+export const EditSafety = ({ currentGear, updateGear }) => {
     const [toggle, setToggle] = useState(false)
-
     const toggleList = () => toggle !== true ? true : false
     const toggled = toggleList()
